finish abstracting fbos
