+++ include/shadow_map.h
@@ -1,0 +1,8 @@
+#ifndef SHADOWMAP_H
+#define SHADOWMAP_H
+
+#include <fbo.h>
+
+FBO create_shadow_map(uint width, uint height);
+
+#endif
